naming cleanup
